requirements.txt
fastapi
uvicorn
pandas
matplotlib
openai
python-dotenv
pydantic
seaborn